feat : add fetching posts function
