+++ app/posts/page.tsx
@@ -1,5 +1,25 @@
 "use client";
 
+import { useState, useEffect } from "react";
+
 export default function Post() {
+  const [data, setData] = useState("");
+
+  useEffect(() => {
+    const fetchData = async () => {
+      try {
+        const response = await fetch("/api/post/get-all-posts");
+        if (!response.ok) {
+          throw new Error("Failed to fetch data");
+        }
+        const result = await response.json();
+        setData(result);
+      } catch (error) {
+        console.error("Error fetching data:", error);
+      }
+    };
+
+    fetchData();
+  }, []);
   return <div></div>;
 }
